chore: Add Fibonacci DAG and task
The code changes add a new file `fib.py` which contains a DAG and a task for calculating the nth Fibonacci number. This will be useful for performing Fibonacci calculations in Airflow workflows.

Note: The commit message format follows the convention of starting with a type (chore) followed by a brief description of the changes.